Fixes https://github.com/mapbox/mapbox-gl-native/issues/7599

+++ src/shaders/symbol_sdf.fragment.glsl
@@ -25,7 +25,7 @@ void main() {
     #pragma mapbox: initialize lowp float halo_blur
 
     lowp vec4 color = fill_color;
-    lowp float gamma = EDGE_GAMMA / u_gamma_scale;
+    highp float gamma = EDGE_GAMMA / u_gamma_scale;
     lowp float buff = (256.0 - 64.0) / 256.0;
     if (u_is_halo) {
         color = halo_color;
